fixed missing call to AbstractDiscreteLight::updateModelToWorldMatrix() in derived classes ctor added tests for the value of a DirectionalLight world space direction now uncluding glm/glx/transform.hpp to have math::tranlate/rotate/scale methods that can take just a math::vec3

+++ framework/src/minko/data/Container.cpp
@@ -124,7 +124,7 @@ Container::getProviderByPropertyName(const std::string& propertyName) const
             if (collection->name() == collectionName)
             {
                 auto pos2 = propertyName.find_first_of("]");
-                auto index = std::stoi(propertyName.substr(pos + 1, pos2 - pos - 1));
+                auto index = (uint)std::stoi(propertyName.substr(pos + 1, pos2 - pos - 1));
 
                 return std::pair<Provider::Ptr, std::string>(
                     index < collection->items().size() ? collection->items()[index] : nullptr,
